Float

+++ docs/js/settings.js
@@ -6,7 +6,7 @@ function getSettings() {
         endDate: Date.parse(document.getElementById("endDate").value),
         budget: parseInt(document.getElementById("budget").value),
         count: parseInt(document.getElementById("count").value),
-        margin: parseInt(document.getElementById("margin").value),
+        margin: parseFloat(document.getElementById("margin").value),
         multiplier: parseFloat(document.getElementById("multiplier").value),
         sellProfit: parseInt(document.getElementById("sellProfit").value),
         buyFall: parseInt(document.getElementById("buyFall").value),
